Added openfed-upgrade composer command

+++ scripts/composer/OpenfedUpdate.php
@@ -19,6 +19,16 @@ class OpenfedUpdate {
    */
   protected static $openfed8Zip = 'https://github.com/openfed/openfed8-project/archive/refs/tags/';
 
+    /**
+   * @var string
+   */
+  protected static $openfedRepo = 'https://github.com/openfed/openfed-project';
+
+  /**
+   * @var string
+   */
+  protected static $openfedZip = 'https://github.com/openfed/openfed-project/archive/refs/tags/';
+
   /**
    * @var string
    */
@@ -30,7 +40,7 @@ class OpenfedUpdate {
    * @param \Composer\Script\Event $event
    */
   public static function update(Event $event) {
-    self::_setLatestOpenfedVersion();
+    self::_setLatestOpenfedVersion(self::$openfed8Repo);
 
     // Check if there's a new Openfed version and update if so.
     if (self::_newVersionExists()) {
@@ -47,7 +57,7 @@ class OpenfedUpdate {
       $response = \Drupal::httpClient()->get($url, ['sink' => $zip_resource]);
 
       if (!$response) {
-        echo "Error :- ";
+        echo "Error :- Cannot connect.";
       }
 
       $zip = new ZipArchive;
@@ -65,6 +75,53 @@ class OpenfedUpdate {
       unlink($extractPath . DIRECTORY_SEPARATOR . 'openfed8-project-' . self::$latestOpenfedVersion . DIRECTORY_SEPARATOR . 'README.md');
 
       self::_recurseCopy($extractPath . DIRECTORY_SEPARATOR . 'openfed8-project-' . self::$latestOpenfedVersion, '.');
+      self::_deleteDirectory($extractPath);
+
+      echo "---- Files updated. You still have to check your composer.json manually.\n\n";
+    }
+  }
+
+  /**
+   * Upgrade current Openfed to the latest major version.
+   *
+   * @param \Composer\Script\Event $event
+   */
+  public static function upgrade(Event $event) {
+    self::_setLatestOpenfedVersion(self::$openfedRepo);
+
+    // Check if there's a new Openfed version and update if so.
+    if (self::_newVersionExists()) {
+      echo "\n\n---- Project files will be updated to Openfed version " . self::$latestOpenfedVersion . "\n";
+
+      $url = self::$openfedZip . self::$latestOpenfedVersion . '.zip';
+      $zipFile = self::$latestOpenfedVersion . '.zip';
+      $extractPath = self::$latestOpenfedVersion;
+
+      $zip_resource = fopen($zipFile, "w");
+
+      self::_initDrupalContainer();
+      /** @var GuzzleHttp\Psr\Response $response */
+      $response = \Drupal::httpClient()->get($url, ['sink' => $zip_resource]);
+
+      if (!$response) {
+        echo "Error :- Cannot connect.";
+      }
+
+      $zip = new ZipArchive;
+      if ($zip->open($zipFile) != "true") {
+        throw new \ErrorException("Error :- Unable to open the Zip File.");
+      }
+
+      $zip->extractTo($extractPath);
+      $zip->close();
+
+      unlink($zipFile);
+      unlink('./composer.libraries.json');
+      unlink($extractPath . DIRECTORY_SEPARATOR . 'openfed-project-' . self::$latestOpenfedVersion . DIRECTORY_SEPARATOR . 'composer.json');
+      unlink($extractPath . DIRECTORY_SEPARATOR . 'openfed-project-' . self::$latestOpenfedVersion . DIRECTORY_SEPARATOR . '.gitignore');
+      unlink($extractPath . DIRECTORY_SEPARATOR . 'openfed-project-' . self::$latestOpenfedVersion . DIRECTORY_SEPARATOR . 'README.md');
+
+      self::_recurseCopy($extractPath . DIRECTORY_SEPARATOR . 'openfed-project-' . self::$latestOpenfedVersion, '.');
       self::_deleteDirectory($extractPath);
 
       echo "---- Files updated. You still have to check your composer.json manually.\n\n";
@@ -130,7 +187,7 @@ class OpenfedUpdate {
    */
   private static function _newVersionExists() {
     $composer_openfed = json_decode(file_get_contents('composer.openfed.json'), TRUE);
-    $current_version = $composer_openfed['require']['openfed/openfed8'];
+    $current_version = isset($composer_openfed['require']['openfed/openfed8']) ? $composer_openfed['require']['openfed/openfed8'] : $composer_openfed['require']['openfed/openfed'];
 
     // If current version is dev, we don't need to check if there's a newer
     // version.
@@ -144,8 +201,8 @@ class OpenfedUpdate {
   /**
    * Set the latest Openfed8 version variable.
    */
-  private static function _setLatestOpenfedVersion() {
-    $latest_openfed_version = explode("\n", trim(shell_exec("git -c 'versionsort.suffix=-' ls-remote --tags --sort='-v:refname' " . self::$openfed8Repo . " | cut --delimiter='/' --fields=3 | grep -v -")));
+  private static function _setLatestOpenfedVersion($repo) {
+    $latest_openfed_version = explode("\n", trim(shell_exec("git -c 'versionsort.suffix=-' ls-remote --tags --sort='-v:refname' " . $repo . " | cut --delimiter='/' --fields=3 | grep -v -")));
     self::$latestOpenfedVersion = $latest_openfed_version[0];
   }
